requirements.txt
dill==0.3.5.1
Flask==2.1.2
gunicorn==20.1.0
ipython
Jinja2==3.1.2
joblib==1.1.0
matplotlib==3.5.2
matplotlib-inline==0.1.3
moves==0.1
numpy
pandas
pickleshare==0.7.5
Pillow==9.1.1
tensorflow-cpu==2.9.1
urllib3==1.26.9
waitress == 2.1.2